Update patch set 1
Patch Set 1:

I'm thinking this one could get merged now to save time, there won't be any changes to the omap4xxx repo with 5.0. The other two commits probably wouldn't merge clean with 5.0 as-is, and I don't see a point to them being merged into 4.4 this late in the cycle.

Patch-set: 1